requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
yfinance
scipy
requests
python-dateutil